Making explicit conversion form double to int  (worksize -> iWorksize) to avoid compiler warnings.

+++ ugbase/lib_algebra/small_algebra/lapack/lapack_invert.h
@@ -23,7 +23,7 @@ bool InvertNdyn(DenseMatrix<T> &mat)
 	double worksize; int iWorksize = -1;
 	info = getri(mat.num_rows(), &mat(0,0), mat.num_rows(), &interchange[0], &worksize, iWorksize);
 	//UG_ASSERT(info == 0, "");
-	iWorksize = worksize;
+	iWorksize = (int)worksize;
 
 	std::vector<double> work(iWorksize);
 
@@ -48,7 +48,7 @@ bool Invert(DenseMatrix<FixedArray2<T, TUnknowns, TUnknowns> > &mat)
 	double worksize; int iWorksize = -1;
 	info = getri(mat.num_rows(), &mat(0,0), mat.num_rows(), interchange, &worksize, iWorksize);
 	UG_ASSERT(info == 0, "");
-	iWorksize = worksize;
+	iWorksize = (int)worksize;
 
 	std::vector<double> work;
 	work.resize(iWorksize);
